+++ priv/repo/migrations/20230904150851_create_metric_entries.exs
@@ -4,7 +4,7 @@ defmodule VrpBenchmarkDataService.Repo.Migrations.CreateMetricEntries do
   def change do
     create table(:metric_entries, primary_key: false) do
       add(:id, :binary_id, primary_key: true)
-      add(:travel_time, :integer)
+      add(:travel_time, :float)
       add(:problem_id, references(:problems, on_delete: :delete_all, type: :binary_id))
       add(:from_node_id, references(:nodes, on_delete: :nothing, type: :binary_id))
       add(:to_node_id, references(:nodes, on_delete: :nothing, type: :binary_id))
